drunkstar: the kernel may not be ready when postprocess is called (do a better cleanup later)

+++ drunkstar/postprocess.py
@@ -16,7 +16,8 @@ def copy_kernel():
     if not os.path.exists(boot):
         os.mkdir(boot)
     kernel = os.path.join(os.path.dirname(sys.argv[1]), "images", "uImage")
-    shutil.copy2(kernel, boot)
+    if os.path.exists(kernel):
+        shutil.copy2(kernel, boot)
 
 if __name__ == "__main__":
     copy_skeleton()
